+++ package.json
@@ -7,6 +7,7 @@
   "license": "MIT",
   "dependencies": {
     "canvas": "^1.2.1",
+    "cron": "^1.0.9",
     "node-twitter-api": "^1.5.0",
     "twit": "^1.1.20"
   }
